requirements.txt
# =============================================================
# NSE Stock Market Pipeline — requirements.txt
# Pin major.minor to prevent silent API-breaking upgrades.
# =============================================================

# --- Data extraction ----------------------------------------
yfinance==0.2.40           # NSE tick & fundamental data
requests==2.31.0           # HTTP for news API
newsapi-python==0.2.7      # NewsAPI client (Rule 6)

# --- Data processing ----------------------------------------
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0            # Parquet / Arrow serialisation

# --- Validation / contracts ---------------------------------
pydantic==2.7.1            # Sprint 2 schema contracts
pydantic-settings==2.3.1   # env-var injection

# --- Database -----------------------------------------------
psycopg2-binary==2.9.9     # Postgres driver
SQLAlchemy==2.0.30         # ORM + Core (Sprint 2 load.py)

# --- ChromaDB (Rule 2) -------------------------------------
chromadb==0.5.5            # Must match docker image tag

# --- Embeddings / LLM (Rules 1, 7) -------------------------
huggingface-hub==0.23.4    # HF inference API (Rule 1 budget)
transformers==4.41.2       # BART summariser (Sprint 5)
torch==2.3.1               # BART runtime — CPU-only build
sentencepiece==0.2.0       # tokeniser support
sentence-transformers==3.0.1  # all-MiniLM-L6-v2 (Rule 7)

# --- PDF extraction (Rule 3) --------------------------------
pdfminer.six==20221105     # primary PDF parser
pytesseract==0.3.10        # OCR fallback (Rule 3)
Pillow==10.3.0             # image support for OCR

# --- Airflow extras -----------------------------------------
apache-airflow==2.8.4
apache-airflow-providers-postgres==5.10.2

# --- Frontend -----------------------------------------------
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0

# --- Testing / CI -------------------------------------------
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
responses==0.25.3          # mock HTTP in tests

# --- Utilities ----------------------------------------------
python-dotenv==1.0.1       # .env loader
tenacity==8.4.1            # retry logic (Rules 1, 6)
loguru==0.7.2              # structured logging
hashlib2==1.0.0            # (stdlib hashlib used directly)
tqdm==4.66.4               # progress bars
